Simplify system
