+++ resources/ts/Pages/Integrations/Detail.tsx
@@ -23,7 +23,6 @@ import { PricingPlanProvider } from "../../Context/PricingPlan";
 import { useIsMobile } from "../../hooks/useIsMobile";
 import { CouponInfoProvider } from "../../Context/CouponInfo";
 import type { Integration } from "../../types/Integration";
-import { usePropsPolling } from "../../hooks/usePropsPolling";
 
 type Props = {
   integration: Integration;
@@ -72,11 +71,6 @@ const Detail = ({ integration, email, subscriptions }: Props) => {
     setIsKeepChangesDialogVisible(false);
     setOriginalVisitingUrl("");
   };
-
-  usePropsPolling([
-    "integration.authClients.length",
-    "integration.legacyAuthConsumers.length",
-  ]);
 
   useEffect(() => {
     const cleanUp = router.on("before", (e) => {
